"bunches of chnages added black keys"

+++ src/com/model/NotePlay.java
@@ -2,25 +2,25 @@ package com.model;
 
 public class NotePlay {
 	private int duration;
-	private int note;
+	private double note;
 	private int counter;
 
 	public NotePlay() {
 		super();
 		this.duration = 0;
-		this.note = 0;
+		this.note = 0.0;
 		this.counter = 0;
 	}
 
 	// counter will index until hits duration
-	public NotePlay(int duration, int note) {
+	public NotePlay(int duration, double note) {
 		super();
 		this.duration = duration;
 		this.note = note;
 		this.counter = 0;
 	}
 
-	public void setValues(int duration, int note) {
+	public void setValues(int duration, double note) {
 		this.duration = duration;
 		this.note = note;
 	}
@@ -41,7 +41,7 @@ public class NotePlay {
 		this.duration = duration;
 	}
 
-	public int getNote() {
+	public double getNote() {
 		return note;
 	}
